+++ devel/sync-sei-devnet/start.sh
@@ -23,6 +23,13 @@ main() {
   set -e
 
   if [[ $clean == "true" ]]; then
+    printf "Are you sure you want to clean the data directory? [y/N] "
+    read -r answer
+    if [[ $answer != "y" ]]; then
+      echo "Aborting"
+      exit 0
+    fi
+
     rm -rf "$fh_data_dir" &> /dev/null || true
   fi
